requirements.txt
pySerial
requests
Flask